suseconds_t is int, not c_long

+++ druntime/src/core/sys/posix/sys/types.d
@@ -391,7 +391,7 @@ else version (NetBSD)
     alias clock_t = uint; // unsigned int
     alias id_t = long;
     alias key_t = c_long;
-    alias suseconds_t = c_long;
+    alias suseconds_t = int; // int
     alias useconds_t = uint;
 }
 else version (OpenBSD)
